update mcluster and container code

+++ cloudplatform-webportal-data-model/src/main/java/com/letv/portal/enumeration/ByteEnum.java
@@ -2,6 +2,6 @@ package com.letv.portal.enumeration;
 
 public interface ByteEnum {
 	 
-	byte getValue();
+	Integer getValue();
 
 }
